refactoring the command loader by adding module command-loader.js
